+++ src/applications/vaos/tests/components/calendar/CalendarCell.unit.spec.jsx
@@ -9,13 +9,14 @@ describe('VAOS <CalendarCell>', () => {
     const tree = shallow(
       <CalendarCell
         date="2018-10-04"
+        currentlySelectedDate="2018-10-04"
         isCurrentlySelected
         inSelectedArray
         disabled={false}
       />,
     );
     const cell = tree.find('button#date-cell-2018-10-04');
-    expect(cell.length).to.equal(1);
+    expect(cell.exists()).to.be.true;
     expect(tree.find('.vaos-calendar__cell-current').length).to.equal(1);
     expect(tree.find('.vaos-calendar__cell-selected').length).to.equal(1);
     expect(cell.text()).to.equal('4');
